+++ ext/dbctrl.py
@@ -4,4 +4,4 @@ def close_connection(connection):
     with closing(connection) as connection:
         with closing(connection.cursor()) as cursor:
             rows = cursor.execute("SELECT 1").fetchall()
-            print(rows)
+            return rows