+++ samples/table_shit.py
@@ -1,3 +1,0 @@
-def gettable():
-	import table
-	shit = table.shutty()