+++ samples/table_test.py
@@ -1,6 +1,0 @@
-def gettable():
-	import table
-	test = table.modernTable()
-	c1 = test.new_column("hey")
-	c2 = test.new_column("bro")
-	test.insert("whatsup","woohoo")