+++ samples/table_tmp.py
@@ -1,4 +1,0 @@
-def gettable():
-	import table
-	tmp = table.modernTable()
-	btn = tmp.new_column("submit")